requirements.txt
ebi-eva-common-pyutils>=0.5
pyyaml
requests
retry
